Added an Asset postprocessor to check for changes to asmdef or csproj files. This will allow us to regen or clean at appropriate times.
